pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibemem"
version = "0.1.0"
description = "Universal memory layer for AI coding tools"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "vibemem contributors" }
]
dependencies = [
    "click>=8.0",
    "tiktoken>=0.5",
    "pyyaml>=6.0",
    "rich>=13.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
vibemem = "vibemem.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["vibemem"]
